fix: changed int back to double

+++ src/common/CircleCollider.cpp
@@ -8,7 +8,7 @@
 
 #include "../headers/CircleCollider.h"
 
-CircleCollider::CircleCollider(int x, int y, int angle, int radius) {
+CircleCollider::CircleCollider(double x, double y, double angle, double radius) {
     this->_position = new Position(x, y, angle);
     this->_colliderType = ect_circle_collider;
     this->_radius = radius;
@@ -18,10 +18,10 @@ bool CircleCollider::CheckCollision(ICollider * collider) {
     return false; // TODO
 }
 
-int CircleCollider::GetRadius() {
+double CircleCollider::GetRadius() {
     return this->_radius;
 }
 
-void CircleCollider::SetRadius(int radius) {
+void CircleCollider::SetRadius(double radius) {
     this->_radius = radius;
 }
